@@ -1,45 +1,46 @@
 {
-  "name": "create-react-personal",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.tsx",
-  "babel": {
-    "presets": [
-      "@babel/preset-env",
-      "@babel/preset-react"
-    ]
-  },
-  "scripts": {
-    "start": "webpack-dev-server --open",
-    "build": "webpack --mode production"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@types/react-router-dom": "^5.1.4",
-    "node-sass": "^4.14.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.1.2"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "@babel/preset-env": "^7.9.5",
-    "@babel/preset-react": "^7.9.4",
-    "@types/react": "^16.9.34",
-    "@types/react-dom": "^16.9.6",
-    "babel-loader": "^8.1.0",
-    "css-loader": "^3.5.2",
-    "html-webpack-plugin": "^4.2.0",
-    "sass": "^1.26.3",
-    "sass-loader": "^8.0.2",
-    "source-map-loader": "^0.2.4",
-    "style-loader": "^1.1.4",
-    "ts-loader": "^7.0.1",
-    "typescript": "^3.8.3",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
-  }
+	"name": "create-react-personal",
+	"type": "module",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.tsx",
+	"babel": {
+		"presets": [
+			"@babel/preset-env",
+			"@babel/preset-react"
+		]
+	},
+	"scripts": {
+		"start": "webpack serve",
+		"build": "webpack --mode production"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"@types/react-router-dom": "^5.1.7",
+		"node-sass": "^5.0.0",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2",
+		"react-router-dom": "^5.2.0"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.13.15",
+		"@babel/preset-env": "^7.13.15",
+		"@babel/preset-react": "^7.13.13",
+		"@types/react": "^17.0.3",
+		"@types/react-dom": "^17.0.3",
+		"babel-loader": "^8.2.2",
+		"css-loader": "^5.2.1",
+		"html-webpack-plugin": "^5.3.1",
+		"sass": "^1.32.8",
+		"sass-loader": "^11.0.1",
+		"source-map-loader": "^2.0.1",
+		"style-loader": "^2.0.0",
+		"ts-loader": "^8.1.0",
+		"typescript": "^4.2.4",
+		"webpack": "^5.37.0",
+		"webpack-cli": "^4.7.0",
+		"webpack-dev-server": "^3.11.2"
+	}
 }
