@@ -1,39 +1,41 @@
-var path = require("path");
-var HtmlWebpackPlugin = require("html-webpack-plugin");
+import path from "path";
+import HtmlWebpackPlugin from "html-webpack-plugin";
 
-module.exports = {
-  entry: "./src/index.tsx",
-  output: {
-    path: path.resolve(__dirname, "dist"),
-    filename: "index_bundle.js",
-  },
-  resolve: {
-    extensions: ["*", ".ts", ".tsx", ".js", ".json"],
-  },
-  module: {
-    rules: [
-      {
-        test: /\.tsx?$/,
-        exclude: /node_modules/,
-        loader: "ts-loader",
-      },
-      {
-        enforce: "pre",
-        test: /\.js$/,
-        loader: "source-map-loader",
-      },
-      { test: /\.(js)$/, use: "babel-loader" },
-      { test: /\.css$/, use: ["style-loader", "css-loader"] },
-      {
-        test: /\.(sa|sc|c)ss$/,
-        use: ["style-loader", "css-loader", "sass-loader"],
-      },
-    ],
-  },
-  mode: "development",
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: "./src/assets/index.html",
-    }),
-  ],
+const __dirname = path.resolve();
+
+export default {
+	entry: "./src/index.tsx",
+	output: {
+		path: path.resolve(__dirname, "dist"),
+		filename: "index_bundle.js",
+	},
+	resolve: {
+		extensions: ["*", ".ts", ".tsx", ".js", ".json"],
+	},
+	module: {
+		rules: [
+			{
+				test: /\.tsx?$/,
+				exclude: /node_modules/,
+				loader: "ts-loader",
+			},
+			{
+				enforce: "pre",
+				test: /\.js$/,
+				loader: "source-map-loader",
+			},
+			{ test: /\.(js)$/, use: "babel-loader" },
+			{ test: /\.css$/, use: ["style-loader", "css-loader"] },
+			{
+				test: /\.(sa|sc|c)ss$/,
+				use: ["style-loader", "css-loader", "sass-loader"],
+			},
+		],
+	},
+	mode: "development",
+	plugins: [
+		new HtmlWebpackPlugin({
+			template: "./src/assets/index.html",
+		}),
+	],
 };
